submit

+++ 4.일차원 배열/1546.py
@@ -5,7 +5,7 @@ data=list(map(int,sys.stdin.readline().split()))
 b=max(data)
 
 for y in range(0,len(data)):
-    data[y]=int((data[y]*100)/b)
+    data[y]=(data[y]*100)/b
     
 for y in range(0,len(data)):
     sum+=data[y]
